Cleaned up Linear regression model  with vtreat
The performance seems very close to the best measure.
The things I tried were as follows
1. log transformation
2. remove outliers
3. impact coding of high arity categorical variables only: This did
about as good as the best regression model I had, but not better

Added code mostly in utils.R to preprocess data. The data processing
flow is:

1. Recoding: Recodes all the char things to integers.
2. Imputing some common sense features: These are arbitrary chosen
features.
3. Convert some of the chosen features to factors. We do this only for
using vtreat because most factors are really large.
4. Make a cross frame and generate treatments for training.. Do not add
the treatments to the data set yet. We need to log transform and exclude
some features first.
5. Log transform some features in the data.
6. Exclude some features from both train and test data sets. Take care
to exclude the features that were also treated using vtreat.
7. Prune the features that are largely missing. Again this is why vtreat
should be used before this step.
8. Merge treated data frame with training data set. Not the testing data
set.
9. Omit NAs.
10 Return the processed data and treatments.